refactor : cors 설정 추가 및 InterceptorConfig와 병합
