--- src/main/java/com/flab/oasis/config/InterceptorConfig.java
@@ -1,20 +1,0 @@
-package com.flab.oasis.config;
-
-import com.flab.oasis.interceptor.TokenInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class InterceptorConfig implements WebMvcConfigurer {
-    private final TokenInterceptor tokenInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(tokenInterceptor)
-                .addPathPatterns("/*")
-                .excludePathPatterns("/api/auth/refresh");
-    }
-}
